requirements.txt
-i https://pypi.org/simple
pyjwt==1.7.1
pyyaml==5.1
